+++ lib/src/presets.dart
@@ -1,3 +1,4 @@
+import 'package:flutter_emoji_feedback/flutter_emoji_feedback.dart';
 import 'package:flutter_emoji_feedback/gen/assets.gen.dart';
 
 import 'models/emoji_model.dart';
@@ -26,6 +27,90 @@ const List<EmojiModel> classicEmojiPreset = [
   ),
   EmojiModel(
     src: Assets.classicAwesome,
+    label: 'Awesome',
+    package: 'flutter_emoji_feedback',
+  ),
+];
+
+const List<EmojiModel> handDrawnEmojiPreset = [
+  EmojiModel(
+    src: Assets.hdTerrible,
+    label: 'Terrible',
+    package: 'flutter_emoji_feedback',
+  ),
+  EmojiModel(
+    src: Assets.hdBad,
+    label: 'Bad',
+    package: 'flutter_emoji_feedback',
+  ),
+  EmojiModel(
+    src: Assets.hdGood,
+    label: 'Good',
+    package: 'flutter_emoji_feedback',
+  ),
+  EmojiModel(
+    src: Assets.hdVeryGood,
+    label: 'Very good',
+    package: 'flutter_emoji_feedback',
+  ),
+  EmojiModel(
+    src: Assets.hdAwesome,
+    label: 'Awesome',
+    package: 'flutter_emoji_feedback',
+  ),
+];
+
+const List<EmojiModel> flatEmojiPreset = [
+  EmojiModel(
+    src: Assets.flatTerrible,
+    label: 'Terrible',
+    package: 'flutter_emoji_feedback',
+  ),
+  EmojiModel(
+    src: Assets.flatBad,
+    label: 'Bad',
+    package: 'flutter_emoji_feedback',
+  ),
+  EmojiModel(
+    src: Assets.flatGood,
+    label: 'Good',
+    package: 'flutter_emoji_feedback',
+  ),
+  EmojiModel(
+    src: Assets.flatVeryGood,
+    label: 'Very good',
+    package: 'flutter_emoji_feedback',
+  ),
+  EmojiModel(
+    src: Assets.flatAwesome,
+    label: 'Awesome',
+    package: 'flutter_emoji_feedback',
+  ),
+];
+
+const List<EmojiModel> threeDEmojiPreset = [
+  EmojiModel(
+    src: Assets.threeDTerrible,
+    label: 'Terrible',
+    package: 'flutter_emoji_feedback',
+  ),
+  EmojiModel(
+    src: Assets.threeDBad,
+    label: 'Bad',
+    package: 'flutter_emoji_feedback',
+  ),
+  EmojiModel(
+    src: Assets.threeDGood,
+    label: 'Good',
+    package: 'flutter_emoji_feedback',
+  ),
+  EmojiModel(
+    src: Assets.threeDVeryGood,
+    label: 'Very good',
+    package: 'flutter_emoji_feedback',
+  ),
+  EmojiModel(
+    src: Assets.threeDAwesome,
     label: 'Awesome',
     package: 'flutter_emoji_feedback',
   ),